reorganize code and handle context done
