+++ client/internal/routemanager/firewall.go
@@ -1,0 +1,8 @@
+package routemanager
+
+type firewallManager interface {
+	RestoreOrCreateContainers() error
+	InsertRoutingRules(pair RouterPair) error
+	RemoveRoutingRules(pair RouterPair) error
+	CleanRoutingRules()
+}
