+++ client/internal/routemanager/manager.go
@@ -18,7 +18,7 @@ type Manager struct {
 	stop           context.CancelFunc
 	mux            sync.Mutex
 	clientRoutes   map[string]*route.Route
-	clientNetworks map[netip.Prefix]*clientNetwork
+	clientNetworks map[string]*clientNetwork
 	serverRoutes   map[string]*route.Route
 	serverRouter   *serverRouter
 	statusRecorder *status.Status
@@ -47,20 +47,12 @@ type serverRouter struct {
 	firewall                 firewallManager
 }
 
-type firewallManager interface {
-	RestoreOrCreateContainers() error
-	InsertRoutingRules(pair RouterPair) error
-	RemoveRoutingRules(pair RouterPair) error
-}
 type RouterPair struct {
 	ID          string
 	source      string
 	destination string
 	masquerade  bool
 }
-
-// DefaultServerCheckInterval default route worker check interval 5s
-const DefaultServerCheckInterval time.Duration = 15000000000
 
 type routerPeerStatus struct {
 	connected bool
@@ -74,7 +66,7 @@ func NewManager(ctx context.Context, pubKey string, wgInterface *iface.WGIface, 
 		ctx:            mCTX,
 		stop:           cancel,
 		clientRoutes:   make(map[string]*route.Route),
-		clientNetworks: make(map[netip.Prefix]*clientNetwork),
+		clientNetworks: make(map[string]*clientNetwork),
 		serverRoutes:   make(map[string]*route.Route),
 		serverRouter: &serverRouter{
 			routes:                   make(map[string]*route.Route),
@@ -89,187 +81,215 @@ func NewManager(ctx context.Context, pubKey string, wgInterface *iface.WGIface, 
 
 func (m *Manager) Stop() {
 	m.stop()
+	m.serverRouter.firewall.CleanRoutingRules()
 }
 
 func (m *Manager) UpdateRoutes(newRoutes []*route.Route) error {
-	m.mux.Lock()
-	defer m.mux.Unlock()
-	clientRoutesToRemove := make([]string, 0)
-	clientRoutesToUpdate := make([]string, 0)
-	clientRoutesToAdd := make([]string, 0)
-	serverRoutesToRemove := make([]string, 0)
-	serverRoutesToUpdate := make([]string, 0)
-	serverRoutesToAdd := make([]string, 0)
-	newClientRoutesMap := make(map[string]*route.Route)
-	newServerRoutesMap := make(map[string]*route.Route)
-	for _, route := range newRoutes {
-		if route.Peer == m.pubKey && runtime.GOOS == "linux" {
-			newServerRoutesMap[route.ID] = route
-			_, found := m.serverRoutes[route.ID]
-			if !found {
-				serverRoutesToAdd = append(serverRoutesToAdd, route.ID)
+	select {
+	case <-m.ctx.Done():
+		log.Infof("not updating routes as context is closed")
+		return m.ctx.Err()
+	default:
+		m.mux.Lock()
+		defer m.mux.Unlock()
+		clientRoutesToRemove := make([]string, 0)
+		clientRoutesToUpdate := make([]string, 0)
+		clientRoutesToAdd := make([]string, 0)
+		serverRoutesToRemove := make([]string, 0)
+		serverRoutesToUpdate := make([]string, 0)
+		serverRoutesToAdd := make([]string, 0)
+		newClientRoutesMap := make(map[string]*route.Route)
+		newServerRoutesMap := make(map[string]*route.Route)
+		for _, newRoute := range newRoutes {
+			// only linux is supported for now
+			if newRoute.Peer == m.pubKey && runtime.GOOS == "linux" {
+				newServerRoutesMap[newRoute.ID] = newRoute
+				_, found := m.serverRoutes[newRoute.ID]
+				if !found {
+					serverRoutesToAdd = append(serverRoutesToAdd, newRoute.ID)
+				}
+			} else {
+				newClientRoutesMap[newRoute.ID] = newRoute
+				_, found := m.clientRoutes[newRoute.ID]
+				if !found {
+					clientRoutesToAdd = append(clientRoutesToAdd, newRoute.ID)
+				}
 			}
-		} else {
-			newClientRoutesMap[route.ID] = route
-			_, found := m.clientRoutes[route.ID]
-			if !found {
-				clientRoutesToAdd = append(clientRoutesToAdd, route.ID)
-			}
-		}
-	}
-
-	if len(newServerRoutesMap) > 0 {
-		err := m.serverRouter.firewall.RestoreOrCreateContainers()
-		if err != nil {
-			// todo
-			log.Fatal(err)
-		}
-	}
-
-	for routeID, _ := range m.clientRoutes {
-		update, found := newClientRoutesMap[routeID]
-		if !found {
-			clientRoutesToRemove = append(clientRoutesToRemove, routeID)
-			continue
 		}
 
-		if !update.IsEqual(m.clientRoutes[routeID]) {
-			clientRoutesToUpdate = append(clientRoutesToUpdate, routeID)
-		}
-	}
-
-	for routeID, _ := range m.serverRoutes {
-		update, found := newServerRoutesMap[routeID]
-		if !found {
-			serverRoutesToRemove = append(serverRoutesToRemove, routeID)
-			continue
-		}
-
-		if !update.IsEqual(m.serverRoutes[routeID]) {
-			serverRoutesToUpdate = append(serverRoutesToUpdate, routeID)
-		}
-	}
-
-	log.Infof("client routes to add %d, remove %d and update %d", len(clientRoutesToAdd), len(clientRoutesToRemove), len(clientRoutesToUpdate))
-
-	for _, routeID := range clientRoutesToRemove {
-		oldRoute := m.clientRoutes[routeID]
-		delete(m.clientRoutes, routeID)
-		m.removeFromClientNetwork(oldRoute)
-	}
-	for _, routeID := range clientRoutesToUpdate {
-		newRoute := newClientRoutesMap[routeID]
-		oldRoute := m.clientRoutes[routeID]
-		m.clientRoutes[routeID] = newRoute
-		if newRoute.Network != oldRoute.Network {
-			m.removeFromClientNetwork(oldRoute)
-		}
-		m.updateClientNetwork(newRoute)
-	}
-	for _, routeID := range clientRoutesToAdd {
-		newRoute := newClientRoutesMap[routeID]
-		m.clientRoutes[routeID] = newRoute
-		m.updateClientNetwork(newRoute)
-	}
-	for id, prefix := range m.clientNetworks {
-		prefix.mux.Lock()
-		if len(prefix.routes) == 0 {
-			log.Debugf("stopping client prefix, %s", prefix.prefix)
-			prefix.stop()
-			delete(m.clientNetworks, id)
-		}
-		prefix.mux.Unlock()
-	}
-
-	log.Infof("client routes added %d, removed %d and updated %d", len(clientRoutesToAdd), len(clientRoutesToRemove), len(clientRoutesToUpdate))
-
-	for _, routeID := range serverRoutesToRemove {
-		oldRoute := m.serverRoutes[routeID]
-		err := m.removeFromServerNetwork(oldRoute)
-		if err != nil {
-			log.Errorf("unable to remove route from server, got: %v", err)
-		}
-		delete(m.serverRoutes, routeID)
-	}
-	for _, routeID := range serverRoutesToUpdate {
-		newRoute := newServerRoutesMap[routeID]
-		oldRoute := m.serverRoutes[routeID]
-
-		var err error
-		if newRoute.Network != oldRoute.Network {
-			err = m.removeFromServerNetwork(oldRoute)
+		if len(newServerRoutesMap) > 0 {
+			err := m.serverRouter.firewall.RestoreOrCreateContainers()
 			if err != nil {
-				log.Errorf("unable to update and remove route %s from server, got: %v", oldRoute.ID, err)
+				// todo
+				log.Fatal(err)
+			}
+		}
+
+		for routeID := range m.clientRoutes {
+			update, found := newClientRoutesMap[routeID]
+			if !found {
+				clientRoutesToRemove = append(clientRoutesToRemove, routeID)
 				continue
 			}
-		}
-		err = m.addToServerNetwork(newRoute)
-		if err != nil {
-			log.Errorf("unable to update and add route %s from server, got: %v", newRoute.ID, err)
-			continue
-		}
-		m.serverRoutes[routeID] = newRoute
-	}
-	for _, routeID := range serverRoutesToAdd {
-		newRoute := newServerRoutesMap[routeID]
-		err := m.addToServerNetwork(newRoute)
-		if err != nil {
-			log.Errorf("unable to add route %s from server, got: %v", newRoute.ID, err)
-			continue
-		}
-		m.serverRoutes[routeID] = newRoute
-	}
 
-	if len(m.serverRoutes) > 0 {
-		enableIPForwarding()
-	}
+			if !update.IsEqual(m.clientRoutes[routeID]) {
+				clientRoutesToUpdate = append(clientRoutesToUpdate, routeID)
+			}
+		}
 
-	log.Infof("server routes added %d, removed %d and updated %d", len(serverRoutesToAdd), len(serverRoutesToRemove), len(serverRoutesToUpdate))
-	return nil
+		for routeID := range m.serverRoutes {
+			update, found := newServerRoutesMap[routeID]
+			if !found {
+				serverRoutesToRemove = append(serverRoutesToRemove, routeID)
+				continue
+			}
+
+			if !update.IsEqual(m.serverRoutes[routeID]) {
+				serverRoutesToUpdate = append(serverRoutesToUpdate, routeID)
+			}
+		}
+
+		log.Infof("client routes to add %d, remove %d and update %d", len(clientRoutesToAdd), len(clientRoutesToRemove), len(clientRoutesToUpdate))
+
+		for _, routeID := range clientRoutesToRemove {
+			oldRoute := m.clientRoutes[routeID]
+			delete(m.clientRoutes, routeID)
+			m.removeFromClientNetwork(oldRoute)
+		}
+		for _, routeID := range clientRoutesToUpdate {
+			newRoute := newClientRoutesMap[routeID]
+			oldRoute := m.clientRoutes[routeID]
+			m.clientRoutes[routeID] = newRoute
+			if newRoute.Network != oldRoute.Network {
+				m.removeFromClientNetwork(oldRoute)
+			}
+			m.updateClientNetwork(newRoute)
+		}
+		for _, routeID := range clientRoutesToAdd {
+			newRoute := newClientRoutesMap[routeID]
+			m.clientRoutes[routeID] = newRoute
+			m.updateClientNetwork(newRoute)
+		}
+		for id, prefix := range m.clientNetworks {
+			prefix.mux.Lock()
+			if len(prefix.routes) == 0 {
+				log.Debugf("stopping client prefix, %s", prefix.prefix)
+				prefix.stop()
+				delete(m.clientNetworks, id)
+			}
+			prefix.mux.Unlock()
+		}
+
+		log.Infof("client routes added %d, removed %d and updated %d", len(clientRoutesToAdd), len(clientRoutesToRemove), len(clientRoutesToUpdate))
+
+		for _, routeID := range serverRoutesToRemove {
+			oldRoute := m.serverRoutes[routeID]
+			err := m.removeFromServerNetwork(oldRoute)
+			if err != nil {
+				log.Errorf("unable to remove route from server, got: %v", err)
+			}
+			delete(m.serverRoutes, routeID)
+		}
+		for _, routeID := range serverRoutesToUpdate {
+			newRoute := newServerRoutesMap[routeID]
+			oldRoute := m.serverRoutes[routeID]
+
+			var err error
+			if newRoute.Network != oldRoute.Network {
+				err = m.removeFromServerNetwork(oldRoute)
+				if err != nil {
+					log.Errorf("unable to update and remove route %s from server, got: %v", oldRoute.ID, err)
+					continue
+				}
+			}
+			err = m.addToServerNetwork(newRoute)
+			if err != nil {
+				log.Errorf("unable to update and add route %s from server, got: %v", newRoute.ID, err)
+				continue
+			}
+			m.serverRoutes[routeID] = newRoute
+		}
+		for _, routeID := range serverRoutesToAdd {
+			newRoute := newServerRoutesMap[routeID]
+			err := m.addToServerNetwork(newRoute)
+			if err != nil {
+				log.Errorf("unable to add route %s from server, got: %v", newRoute.ID, err)
+				continue
+			}
+			m.serverRoutes[routeID] = newRoute
+		}
+
+		log.Infof("server routes added %d, removed %d and updated %d", len(serverRoutesToAdd), len(serverRoutesToRemove), len(serverRoutesToUpdate))
+
+		if len(m.serverRoutes) > 0 {
+			err := enableIPForwarding()
+			if err != nil {
+				return err
+			}
+		}
+
+		return nil
+	}
+}
+
+func getClientNetworkID(input *route.Route) string {
+	return input.NetID + "-" + input.Network.String()
 }
 
 func (m *Manager) removeFromClientNetwork(oldRoute *route.Route) {
-	client, found := m.clientNetworks[oldRoute.Network]
-	if !found {
-		log.Debugf("managed prefix %s not found", oldRoute.Network.String())
+	select {
+	case <-m.ctx.Done():
+		log.Infof("not removing from client network because context is done: %v", m.ctx.Err())
 		return
+	default:
+		client, found := m.clientNetworks[getClientNetworkID(oldRoute)]
+		if !found {
+			log.Debugf("managed prefix %s not found", oldRoute.Network.String())
+			return
+		}
+		client.mux.Lock()
+		delete(client.routes, oldRoute.ID)
+		client.mux.Unlock()
+		client.update <- struct{}{}
 	}
-	client.mux.Lock()
-	delete(client.routes, oldRoute.ID)
-	client.mux.Unlock()
-	client.update <- struct{}{}
 }
 
-func (m *Manager) startClientNetworkWatcher(prefixString string) *clientNetwork {
-	prefix, _ := netip.ParsePrefix(prefixString)
+func (m *Manager) startClientNetworkWatcher(networkRoute *route.Route) *clientNetwork {
 	ctx, cancel := context.WithCancel(m.ctx)
 	client := &clientNetwork{
 		ctx:    ctx,
 		stop:   cancel,
 		routes: make(map[string]*route.Route),
 		update: make(chan struct{}),
-		prefix: prefix,
+		prefix: networkRoute.Network,
 	}
-	m.clientNetworks[prefix] = client
-	go m.watchClientNetworks(prefix)
+	id := getClientNetworkID(networkRoute)
+	m.clientNetworks[id] = client
+	go m.watchClientNetworks(id)
 	return client
 }
 
 func (m *Manager) updateClientNetwork(newRoute *route.Route) {
-	client, found := m.clientNetworks[newRoute.Network]
-	if !found {
-		client = m.startClientNetworkWatcher(newRoute.Network.String())
+	select {
+	case <-m.ctx.Done():
+		log.Infof("not updating client network because context is done: %v", m.ctx.Err())
+		return
+	default:
+		client, found := m.clientNetworks[newRoute.NetID+newRoute.Network.String()]
+		if !found {
+			client = m.startClientNetworkWatcher(newRoute)
+		}
+		client.mux.Lock()
+		client.routes[newRoute.ID] = newRoute
+		client.mux.Unlock()
+		client.update <- struct{}{}
 	}
-	client.mux.Lock()
-	client.routes[newRoute.ID] = newRoute
-	client.mux.Unlock()
-	client.update <- struct{}{}
 }
 
-func (m *Manager) watchClientNetworks(prefix netip.Prefix) {
-	client, prefixFound := m.clientNetworks[prefix]
+func (m *Manager) watchClientNetworks(id string) {
+	client, prefixFound := m.clientNetworks[id]
 	if !prefixFound {
-		log.Errorf("attepmt to watch prefix %s failed. prefix not found in manager map", prefix.String())
+		log.Errorf("attepmt to watch prefix %s failed. prefix not found in manager map", id)
 		return
 	}
 	ticker := time.NewTicker(DefaultClientCheckInterval)
@@ -288,8 +308,6 @@ func (m *Manager) watchClientNetworks(prefix netip.Prefix) {
 	for {
 		select {
 		case <-client.ctx.Done():
-			// close things
-			// remove prefix from route table
 			log.Debugf("stopping routine for prefix %s", client.prefix)
 			client.mux.Lock()
 			err := removeFromRouteTable(client.prefix)
@@ -308,6 +326,8 @@ func (m *Manager) watchClientNetworks(prefix netip.Prefix) {
 					if found {
 						removeErr := m.wgInterface.RemoveAllowedIP(previousChosen.Peer, client.prefix.String())
 						if removeErr != nil {
+							log.Debugf("couldn't remove allowed IP %s removed for peer %s, err: %v",
+								client.prefix, previousChosen.Peer, removeErr)
 							client.mux.Unlock()
 							continue
 						}
@@ -317,6 +337,8 @@ func (m *Manager) watchClientNetworks(prefix netip.Prefix) {
 					chosenRoute := client.routes[chosen]
 					err := m.wgInterface.AddAllowedIP(chosenRoute.Peer, client.prefix.String())
 					if err != nil {
+						log.Errorf("couldn't add allowed IP %s added for peer %s, err: %v",
+							client.prefix, chosenRoute.Peer, err)
 						client.mux.Unlock()
 						continue
 					}
@@ -324,8 +346,10 @@ func (m *Manager) watchClientNetworks(prefix netip.Prefix) {
 					if !found {
 						err = addToRouteTable(client.prefix, m.wgInterface.GetAddress().IP.String())
 						if err != nil {
+							log.Errorf("route %s couldn't be added for peer %s, err: %v",
+								chosenRoute.Network.String(), m.wgInterface.GetAddress().IP.String(), err)
 							client.mux.Unlock()
-							panic(err)
+							continue
 						}
 						log.Debugf("route %s added for peer %s", chosenRoute.Network.String(), m.wgInterface.GetAddress().IP.String())
 					}
@@ -333,7 +357,11 @@ func (m *Manager) watchClientNetworks(prefix netip.Prefix) {
 					log.Debugf("no change on chossen route for prefix %s", client.prefix)
 				}
 			} else {
-				log.Debugf("no route was chosen for prefix %s", client.prefix)
+				var peers []string
+				for _, r := range client.routes {
+					peers = append(peers, r.Peer)
+				}
+				log.Warnf("no route was chosen for prefix %s, no peers from list %s were connected", client.prefix, peers)
 			}
 			client.mux.Unlock()
 		}
@@ -346,18 +374,18 @@ func getBestRoute(routes map[string]*route.Route, routePeerStatuses map[string]r
 
 	for _, r := range routes {
 		tempScore := 0
-		status, found := routePeerStatuses[r.ID]
-		if !found || !status.connected {
+		peerStatus, found := routePeerStatuses[r.ID]
+		if !found || !peerStatus.connected {
 			continue
 		}
 		if r.Metric < route.MaxMetric {
 			metricDiff := route.MaxMetric - r.Metric
 			tempScore = metricDiff * 10
 		}
-		if !status.relayed {
+		if !peerStatus.relayed {
 			tempScore++
 		}
-		if !status.direct {
+		if !peerStatus.direct {
 			tempScore++
 		}
 		if tempScore > chosenScore {
@@ -371,13 +399,13 @@ func getBestRoute(routes map[string]*route.Route, routePeerStatuses map[string]r
 
 func (m *Manager) getRouterPeerStatuses(routes map[string]*route.Route) map[string]routerPeerStatus {
 	routePeerStatuses := make(map[string]routerPeerStatus)
-	for _, route := range routes {
-		peerStatus, err := m.statusRecorder.GetPeer(route.Peer)
+	for _, r := range routes {
+		peerStatus, err := m.statusRecorder.GetPeer(r.Peer)
 		if err != nil {
 			log.Debugf("couldn't fetch peer state: %v", err)
 			continue
 		}
-		routePeerStatuses[route.ID] = routerPeerStatus{
+		routePeerStatuses[r.ID] = routerPeerStatus{
 			connected: peerStatus.ConnStatus == peer.StatusConnected.String(),
 			relayed:   peerStatus.Relayed,
 			direct:    peerStatus.Direct,
@@ -397,23 +425,35 @@ func routeToRouterPair(source string, route *route.Route) RouterPair {
 }
 
 func (m *Manager) removeFromServerNetwork(route *route.Route) error {
-	m.serverRouter.mux.Lock()
-	defer m.serverRouter.mux.Unlock()
-	err := m.serverRouter.firewall.RemoveRoutingRules(routeToRouterPair(m.wgInterface.Address.String(), route))
-	if err != nil {
-		return err
+	select {
+	case <-m.ctx.Done():
+		log.Infof("not removing from server network because context is done")
+		return m.ctx.Err()
+	default:
+		m.serverRouter.mux.Lock()
+		defer m.serverRouter.mux.Unlock()
+		err := m.serverRouter.firewall.RemoveRoutingRules(routeToRouterPair(m.wgInterface.Address.String(), route))
+		if err != nil {
+			return err
+		}
+		delete(m.serverRouter.routes, route.ID)
+		return nil
 	}
-	delete(m.serverRouter.routes, route.ID)
-	return nil
 }
 
 func (m *Manager) addToServerNetwork(route *route.Route) error {
-	m.serverRouter.mux.Lock()
-	defer m.serverRouter.mux.Unlock()
-	err := m.serverRouter.firewall.InsertRoutingRules(routeToRouterPair(m.wgInterface.Address.String(), route))
-	if err != nil {
-		return err
+	select {
+	case <-m.ctx.Done():
+		log.Infof("not adding to server network because context is done")
+		return m.ctx.Err()
+	default:
+		m.serverRouter.mux.Lock()
+		defer m.serverRouter.mux.Unlock()
+		err := m.serverRouter.firewall.InsertRoutingRules(routeToRouterPair(m.wgInterface.Address.String(), route))
+		if err != nil {
+			return err
+		}
+		m.serverRouter.routes[route.ID] = route
+		return nil
 	}
-	m.serverRouter.routes[route.ID] = route
-	return nil
 }
